# contracts/v1/schema/openapi-v1.draft.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "enum": [
          "invalid_request",
          "not_found",
          "scope_mismatch",
          "stale_revision",
          "contract_validation_failed",
          "protocol_integrity_failed",
          "publication_integrity_failed",
          "conflict",
          "job_failed"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "correlation_id": {
            "minLength": 1,
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "recovery_action": {
            "minLength": 1,
            "title": "Recovery Action",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "code",
          "title",
          "detail",
          "recovery_action",
          "correlation_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "schema_version": {
            "const": "fixture/v1",
            "default": "fixture/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Subject": {
        "additionalProperties": false,
        "properties": {
          "age_years": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Age Years"
          },
          "center_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Center Code"
          },
          "center_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Center Name"
          },
          "project_id": {
            "minLength": 1,
            "title": "Project Id",
            "type": "string"
          },
          "revision": {
            "default": 1,
            "minimum": 1,
            "title": "Revision",
            "type": "integer"
          },
          "schema_version": {
            "const": "fixture/v1",
            "default": "fixture/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Sex"
          },
          "subject_code": {
            "minLength": 1,
            "title": "Subject Code",
            "type": "string"
          },
          "subject_id": {
            "minLength": 1,
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "subject_id",
          "subject_code",
          "project_id"
        ],
        "title": "Subject",
        "type": "object"
      },
      "SubjectListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Subject"
            },
            "title": "Items",
            "type": "array"
          },
          "project_id": {
            "minLength": 1,
            "title": "Project Id",
            "type": "string"
          },
          "schema_version": {
            "const": "fixture/v1",
            "default": "fixture/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "items"
        ],
        "title": "SubjectListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "入排审核系统 V2 Stub API",
    "version": "fixture/v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/projects/{project_id}/subjects": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectListResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "请求未完成"
          }
        },
        "summary": "读取项目受试者及分阶段状态"
      }
    },
...
  }
}